Move class Results to own file
